+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/). All notable c
 ## [Unreleased]
 ### Added
 - [#8](https://github.com/OldSneerJaw/couchster/issues/8): Regular expression pattern constraint for document ID
+- [#10](https://github.com/OldSneerJaw/couchster/issues/10): Extended year format in date strings
 
 ## [0.2.0] - 2018-03-08
 ### Added